generalized validation server
