+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/rule/DccValidationRuleApi.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.covidcertificate.validation.core.rule.server
+package de.rki.coronawarnapp.covidcertificate.validation.core.rule
 
 import de.rki.coronawarnapp.covidcertificate.validation.core.rule.DccValidationRule
 import retrofit2.http.GET
